Adjusted pom.xml (newer version of jakarta.xml.bind)
The converter transitively depends on the jakarta.xml.bind-api. It used to depend on version 2.3.3, but now it depends on 4.0.1. Depending on the older version created a conflict when using the converter in the restapi. (in 2.3.3 a packages name is "javax.xml.bind" and in 4.0.1 it is "jakarta.xml.bind")